src/main/java/ch/hesge/algo/L.java
package ch.hesge.algo;

import ch.hesge.algo.model.Company;

import java.util.List;
import java.util.Set;
import java.util.stream.Collectors;

public class L {

    /**
     * En utilisant uniquement l'API Streams,
     * retourner une String dont le nom est la concaténation de toutes les String séparées par " & ".
     *
     * Par exemple :
     *
     * Soit la String "Foo"
     * Soit la String "Bar"
     *
     * La String retournée doit être "Foo & Bar"
     *
     * @param strings List de Strings
     * @return String concaténée
     */
    public String merge(List<String> strings) {
        return strings.stream()
                .collect(Collectors.joining(" & "));
    }
}
